Don't use esc_attr_e() on language code / name
In the templates esc_attr_e() is used to escape the language code. This can potentially causes issues because worpress translates the language code as if it were a word (e.g. Amharic 'am' would be interpreted as in '9 am').

In the templates esc_attr_e() is used to escape the language name. Imho the language name should not be translated in the language switcher.

(this relates to all widget template files)